[bug/9725] More Azure Fixes Resolves port problems and a typo (thanks Evil)

+++ phpBB/common.php
@@ -138,7 +138,7 @@ if (!defined('PHPBB_INSTALLED'))
 	$server_name = (!empty($_SERVER['HTTP_HOST'])) ? strtolower($_SERVER['HTTP_HOST']) : ((!empty($_SERVER['SERVER_NAME'])) ? $_SERVER['SERVER_NAME'] : getenv('SERVER_NAME'));
 	
 	//$server_port = (!empty($_SERVER['SERVER_PORT'])) ? (int) $_SERVER['SERVER_PORT'] : (int) getenv('SERVER_PORT');
-	$server_port_raw = (!empty($_SERVER['HTTP_HOST'])) ? (int) $_SERVER['HTTP_HOST'] : (int) getenv('HTTP_HOST');
+	$server_port_raw = (!empty($_SERVER['HTTP_HOST'])) ? $_SERVER['HTTP_HOST'] : getenv('HTTP_HOST');
 	$server_port = (int) substr($server_port_raw, strpos($server_port_raw, ':'));	
 	$secure = (isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] == 'on') ? 1 : 0;
